+++ CHANGELOG.md
@@ -7,6 +7,10 @@
   useful summary for people upgrading their application, not a replication
   of the commit log.
 
+## Unreleased
+
+* Change id validation on component wrapper ([PR #4584](https://github.com/alphagov/govuk_publishing_components/pull/4584))
+
 ## 50.0.0
 
 * Remove margin top from search component ([PR #4581](https://github.com/alphagov/govuk_publishing_components/pull/4581))